Use coraza.WAFConfig, move the test to where it belongs.
