+++ experimental/rule_observer.go
@@ -1,19 +1,20 @@
 package experimental
 
 import (
+	"github.com/corazawaf/coraza/v3"
 	"github.com/corazawaf/coraza/v3/types"
 )
 
 // wafConfigWithRuleObserver is the private capability interface
 type wafConfigWithRuleObserver interface {
-	WithRuleObserver(func(rule types.RuleMetadata)) any
+	WithRuleObserver(func(rule types.RuleMetadata)) coraza.WAFConfig
 }
 
 // WAFConfigWithRuleObserver applies a rule observer if supported.
 func WAFConfigWithRuleObserver(
-	cfg any,
+	cfg coraza.WAFConfig,
 	observer func(rule types.RuleMetadata),
-) any {
+) coraza.WAFConfig {
 	if c, ok := cfg.(wafConfigWithRuleObserver); ok {
 		return c.WithRuleObserver(observer)
 	}
